refactor: extract RecordingStartPreflightResult.swift from PermissionAndPreflightTypes.swift
Splits the `RecordingStartPreflightResult` enum out of
PermissionAndPreflightTypes.swift so that file holds only pure
permission state/status/recovery types.

Byte-identical move of lines 26-40. Both files import Foundation.

PermissionAndPreflightTypes.swift: 83 → 67 lines. Tests: 667.
